Move domain search provider up the tree
